requirements.txt
redis-py
pillow
torch with cuda support
torchvision
numpy
opencv
joblib
scikit-learn
